Transfer currencyPairs to currencies KT-1
