+++ src/core/models/index.ts
@@ -1,4 +1,3 @@
 export * from './currency-pairs-query-params';
 export * from './exchange-rates';
-export * from './currency-pairs';
 export * from './currency';
